Chore: Split up InformationBoxEE into smaller components
