+++ packages/core/admin/ee/admin/content-manager/pages/EditView/InformationBox/components/AssigneeSelect/index.js
@@ -1,0 +1,1 @@
+export * from './AssigneeSelect';
