+++ packages/core/admin/ee/admin/content-manager/pages/EditView/InformationBox/components/StageSelect/index.js
@@ -1,0 +1,1 @@
+export * from './StageSelect';
