+++ packages/core/admin/ee/admin/content-manager/pages/EditView/InformationBox/constants.js
@@ -1,0 +1,2 @@
+export const STAGE_ATTRIBUTE_NAME = 'strapi_reviewWorkflows_stage';
+export const ASSIGNEE_ATTRIBUTE_NAME = 'strapi_assignee';
